+++ R/Scoper.R
@@ -12,7 +12,9 @@
 #' @section SCOPer:
 #'
 #' \itemize{
-#'   \item  \link{defineClonesScoper}:  Clustering sequences into clonal groups.
+#'   \item  \link{identicalClones}:  Identical clustering-based metho for partitioning Ig sequences into clones..
+#'   \item  \link{hierarchicalClones}:  Hierarchical clustering-based method for partitioning Ig sequences into clones.
+#'   \item  \link{spectralClones}:  Spectral clustering-based method for partitioning Ig sequences into clones.
 #' }
 #'
 #' @name        scoper
@@ -21,6 +23,8 @@
 #' \enumerate{
 #'   \item  Nouri N and Kleinstein SH (2018). A spectral clustering-based method for identifying clones
 #'   from high-throughput B cell repertoire sequencing data. Bioinformatics, 34(13):i341-i349.
+#'   \item  Nouri N and Kleinstein SH (2019). Somatic hypermutation analysis for improved identification 
+#'   of B cell clonal families from next-generation sequencing data. bioRxiv, 10.1101/788620.
 #'   \item Gupta NT, et al. (2017). Hierarchical clustering can identify B cell clones with high 
 #'   confidence in Ig repertoire sequencing data. The Journal of Immunology, 198(6):2489-2499.
 #' }